added negative indexing access

+++ t_stack/t_stack.h
@@ -6,7 +6,7 @@
 /*   By: ikhadem <ikhadem@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/28 14:59:04 by ikhadem           #+#    #+#             */
-/*   Updated: 2021/09/03 13:53:13 by ikhadem          ###   ########.fr       */
+/*   Updated: 2021/09/10 08:10:38 by ikhadem          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ t_stack		*t_stack_copy(t_stack *self);
 size_t		t_stack_size(t_stack *);
 int			t_stack_index_of(t_stack *self, int key);
 int			t_stack_top(t_stack **);
-int			t_stack_index(t_stack *, size_t);
+int			t_stack_index(t_stack *, int);
 int			t_stack_max(t_stack *self);
 int			t_stack_min(t_stack *self);
 int			t_stack_is_sorted(t_stack *self);
